requirements.txt
cachetools==4.2.2
google-auth==1.32.1
google-cloud-bigquery==2.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
socket==1.0.0
